pneumonia-detection-flask/pneumonia-detection-flask/requirements.txt
# Pneumonia Detection - Deep Learning Flask Project
# Python 3.10+

# Web framework
Flask==3.0.0
Werkzeug==3.0.1

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# ML utilities
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Optional: for progress bars during training
tqdm>=4.65.0

# Optional: for dataset download (run: python setup_dataset.py)
kaggle>=1.5.0
